requirements.txt
altair==5.0.1
astroid==2.15.5
attrs==23.1.0
blinker==1.6.2
branca==0.6.0
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.4
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
decorator==5.1.1
dill==0.3.6
filelock==3.12.2
folium==0.14.0
fonttools==4.40.0
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
importlib-metadata==6.7.0
importlib-resources==5.12.0
isort==5.12.0
Jinja2==3.1.2
joblib==1.3.1
jsonschema==4.18.0
jsonschema-specifications==2023.6.1
kiwisolver==1.4.4
lazy-object-proxy==1.9.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
mccabe==0.7.0
mdurl==0.1.2
mpmath==1.3.0
mypy==1.4.1
mypy-extensions==1.0.0
networkx==3.1
numpy==1.25.0
opencv-python==4.8.0.74
packaging==23.1
pandas==2.0.3
pandas-stubs==2.0.2.230605
peewee==3.16.2
Pillow==9.5.0
platformdirs==3.8.1
protobuf==4.23.4
psutil==5.9.5
pyarrow==12.0.1
pydeck==0.8.1b0
Pygments==2.15.1
pylint==2.17.4
Pympler==1.0.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
referencing==0.29.1
requests==2.31.0
rich==13.4.2
rpds-py==0.8.8
scikit-learn==1.3.0
scipy==1.11.1
seaborn==0.12.2
six==1.16.0
smmap==5.0.0
streamlit==1.24.0
streamlit-aggrid==0.3.4.post3
streamlit-folium==0.13.0
sympy==1.12
tenacity==8.2.2
threadpoolctl==3.1.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.8
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
types-peewee==3.16.0.0
types-Pillow==10.0.0.0
types-pytz==2023.3.0.0
types-PyYAML==6.0.12.10
types-tqdm==4.65.0.1
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
ultralytics==8.0.128
urllib3==2.0.3
validators==0.20.0
watchdog==3.0.0
wrapt==1.15.0
zipp==3.15.0
